requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.3.1
kfp==1.8.14
PyYAML>=5.3,<6.0
typing-extensions>=4.5.0
scikit-learn>=1.0.0